backened/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
alembic
psycopg2-binary
python-multipart
pillow
tensorflow==2.12.0
aiofiles
python-jose  # if you add auth later
databases  # optional async db layer
